Update and rename LtRedundantHatTest.java to LtRedundantRhoTest.java
